requirements.txt
requests==0.7.4
mock==0.7.2
nose==1.1.2
sphinx
git+git://github.com/kyleconroy/webtest.git
git+git://github.com/kyleconroy/python-oauth2.git
simplejson
python-dateutil
httplib2
python-oauth2
pytz
djangorestframework
markdown
django-filter
